Cookie Auth in combination with Windows Auth
Windows authentication mechanism is only available to IE and Edge.

Since Safari/Chrome/Firefox don't support authentication over websockets we are now using Cookie Authentication in combination with Windows (Negociate) Authentication.

The user is authenticated over the Negociate Microsoft Windows authentication mechanism, a cookie is created and this cookie is then passed with the websocket to validate the identity of the user.